+++ src/test/container-features/featuresCLICommands.test.ts
@@ -1,6 +1,9 @@
 import { assert } from 'chai';
 import path from 'path';
-import { createPlainLog, LogLevel, makeLog } from '../../spec-utils/log';
+import { createLog } from '../../spec-node/devContainers';
+import { getSermanticVersions } from '../../spec-node/featuresCLI/publishCommandImpl';
+import { getPackageConfig } from '../../spec-node/utils';
+import { createPlainLog, Log, LogLevel, makeLog, mapLogLevel } from '../../spec-utils/log';
 import { isLocalFile, readLocalFile } from '../../spec-utils/pfs';
 import { shellExec } from '../testUtils';
 export const output = makeLog(createPlainLog(text => process.stdout.write(text), () => LogLevel.Trace));
@@ -84,4 +87,62 @@ describe('CLI features subcommands', async function () {
 		assert.strictEqual(json.features.length, 1);
 		assert.isTrue(collectionFileExists);
 	});
+});
+
+describe('features publish subcommand', () => {
+    let output: Log;
+    const disposables: (() => Promise<unknown> | undefined)[] = [];
+    const dispose = async () => {
+        await Promise.all(disposables.map(d => d()));
+    };
+
+    before(async () => {
+
+        const extensionPath = path.join(__dirname, '..', '..');
+        const pkg = await getPackageConfig(extensionPath);
+
+        output = createLog({
+            logLevel: mapLogLevel('trace'),
+            logFormat: 'text',
+            log: (str) => process.stdout.write(str),
+            terminalDimensions: undefined,
+        }, pkg, new Date(), disposables, true);
+    });
+
+    it('should generate correct semantic versions', async () => {
+        // First publish
+        let version = '1.0.0';
+        let publishedVersions: string[] = [];
+        let expectedSemVer = ['1', '1.0', '1.0.0', 'latest'];
+
+        let semanticVersions = getSermanticVersions(version, publishedVersions, output);
+        assert.equal(semanticVersions?.toString(), expectedSemVer.toString());
+
+        // Publish new major version
+        version = '2.0.0';
+        publishedVersions = ['1', '1.0', '1.0.0', 'latest'];
+        expectedSemVer = ['2', '2.0', '2.0.0', 'latest'];
+
+        semanticVersions = getSermanticVersions(version, publishedVersions, output);
+        assert.equal(semanticVersions?.toString(), expectedSemVer.toString());
+
+        // Publish hotfix version
+        version = '1.0.1';
+        publishedVersions = ['1', '1.0', '1.0.0', '2', '2.0', '2.0.0', 'latest'];
+        expectedSemVer = ['1', '1.0', '1.0.1'];
+
+        semanticVersions = getSermanticVersions(version, publishedVersions, output);
+        assert.equal(semanticVersions?.toString(), expectedSemVer.toString());
+
+        // Re-publish version
+        version = '1.0.1';
+        publishedVersions = ['1', '1.0', '1.0.0', '1.0.1', '2', '2.0', '2.0.0', 'latest'];
+
+        semanticVersions = getSermanticVersions(version, publishedVersions, output);
+        assert.isUndefined(semanticVersions);
+    });
+
+    after(async () => {
+        await dispose();
+    });
 });
